retroactive-indexer/requirements.txt
# Editable install with no version control (anchorpy==0.21.0)
-e ../third_party/anchorpy
anchorpy-core==0.2.0
anchorpy-idl==0.1.0
anyio==4.10.0
asttokens==3.0.0
asyncio==4.0.0
attrs==25.3.0
autoflake==2.3.1
base58==2.1.1
based58==0.1.1
black==25.1.0
borsh-construct==0.1.0
certifi==2025.8.3
click==8.2.1
construct==2.10.68
construct-typing==0.5.6
decorator==5.2.1
executing==2.2.1
genpy==2022.1
grpcio==1.74.0
grpcio-tools==1.74.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonalias==0.1.1
markdown-it-py==4.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mypy_extensions==1.1.0
numpy==2.3.2
packaging==25.0
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.4.0
pluggy==1.6.0
prompt_toolkit==3.0.52
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pyflakes==3.4.0
Pygments==2.19.2
pyheck==0.1.5
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-xprocess==1.0.2
python-dotenv==1.1.1
pytools==2025.2.3
rich==14.1.0
setuptools==80.9.0
shellingham==1.5.4
siphash24==1.7
sniffio==1.3.1
solana==0.36.6
solders==0.26.0
stack-data==0.6.3
sumtypes==0.1a6
toml==0.10.2
toolz==0.11.2
traitlets==5.14.3
typer==0.17.3
typing_extensions==4.15.0
wcwidth==0.2.13
websockets==15.0
wheel==0.45.1
